requirements.txt
fastapi
uvicorn
torch
Pillow
git+https://github.com/openai/CLIP.git
tqdm
regex
ftfy
